backend/requirements.txt
asgiref==3.10.0
crispy-bootstrap5==2025.6
Django==5.2.7
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-extensions==4.1
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==12.0.0
sqlparse==0.5.3
